+++ README.md
@@ -31,6 +31,6 @@ For AST parsing we use [tree-sitter](https://github.com/tree-sitter/tree-sitter)
 To work with parsers, do the following:
 
 1. Clone the appropriate tree-sitter parser in `src/parsing/tree-sitter-languages`
-2. Add information about it to [setup_tree_sitter.py](src/parsing/setup_tree_sitter.py)
-3. Add the language to [utils.py](src/parsing/utils.py)
-4. Create a wrapper (see [PythonAstParser](src/parsing/python_ast_parser.py) for example)  
+2. Add information about it to [setup_tree_sitter.py](src/processing/setup_tree_sitter.py)
+3. Add the language to [utils.py](src/processing/utils.py)
+4. Create a wrapper (see [PythonAstParser](src/processing/python_ast_parser.py) for example)
